requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.1.0
Flask-JSON==0.4.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
fonttools==4.55.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pdfkit==1.0.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
gunicorn
Flask-Cors
psycopg2-binary
